Apply suggestions from PR (#23)
