+++ pyproject.toml
@@ -6,14 +6,14 @@ requires-python = "==3.11.15"
 dependencies = [
     "einops>=0.8.2,<1.0.0",
     "jupyter>=1.1.1,<2.0.0",
-    "lpips>=0.1.4",
+    "lpips>=0.1.4,<1.0.0",
     "matplotlib>=3.10.8,<4.0.0",
     "monai>=1.5.2,<2.0.0",
     "mypy>=1.20.0,<2.0.0",
     "nibabel>=5.4.2,<6.0.0",
     "numpy==2.4.4",
     "optuna>=4.8.0,<5.0.0",
-    "pandas>=2.2",
+    "pandas>=2.2,<3.0.0",
     "ruff>=0.15.9,<1.0.0",
     "scikit-image>=0.26.0,<1.0.0",
     "simpleitk>=2.5.3,<3.0.0",
@@ -75,10 +75,6 @@ quote-style = "double"
 [tool.mypy]
 python_version = "3.11"
 strict = true
-
-[[tool.mypy.overrides]]
-module = ["lpips"]
-ignore_missing_imports = true
 
 # Explicitly exclude the notebooks directory from type checking
 exclude = [
